Move the stack code to its own file
